Fixed quoting levels.
This doesn't fix the implementation but at least gets the code parsing
and running. Pushed the quoting up to inc/dec/jnz/etc because they're
the entry points. Doesn't terminate on input file (and hits a point
where it doesn't seem to be doing anything but spinning in the loop).